refactor: move lotto number validation method
- moved Lotto number and perchase count code
  - NumberUtil.java -> Lottos.java
